+++ .eslintrc
@@ -30,7 +30,7 @@
     "import/no-named-as-default": "off",
     "import/no-webpack-loader-syntax": "off",
     "import/no-extraneous-dependencies": "off",
-    "import/no-unresolved": ["error", { "ignore": ["^reakit"] }],
+    "import/no-unresolved": ["error", { "ignore": ["^pulsar-ui"] }],
     "react/require-default-props": "off",
     "react/default-props-match-prop-types": "off",
     "react/no-array-index-key": "off",
