+++ package.json
@@ -16,7 +16,7 @@
     "postcoverage": "opn coverage/lcov-report/index.html",
     "type-check": "tsc --noEmit",
     "lint": "eslint . --ext js,ts,tsx",
-    "dev": "lerna exec --scope reakit -- npm run dev",
+    "dev": "lerna exec --scope @pulsar-ui/core -- npm run dev",
     "build": "lerna exec --scope pulsar-ui* --concurrency 1 -- npm run build",
     "build:clean": "lerna exec --scope pulsar-ui* -- npm run build:clean",
     "website": "lerna exec --scope website -- npm run start",
