--- packages/pulsar-ui-theme-default/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./src";
-export * from "./src";
