--- packages/pulsar-ui-theme-default/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "pulsar-ui-theme-default",
-  "version": "0.1.0",
-  "description": "",
-  "main": "lib/index.js",
-  "module": "es/index.js",
-  "unpkg": "dist/pulsar-ui-theme-default.min.js",
-  "types": "ts",
-  "scripts": {
-    "prebuild": "tsc --emitDeclarationOnly",
-    "build": "rollup -c",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "rollup": "1.2.2",
-    "typescript": "3.2.4"
-  },
-  "peerDependencies": {
-    "pulsar-ui": "*"
-  },
-  "dependencies": {
-    "styled-tools": "1.7.1"
-  }
-}
