--- packages/pulsar-ui-theme-default/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "ts",
-  },
-  "include": [
-    "src",
-    "../../types"
-  ]
-}
