+++ packages/pulsar-ui/.gitignore
@@ -5,3 +5,4 @@
 /Button/
 /Box/
 /Input/
+/Avatar/
