+++ packages/pulsar-ui/tsconfig.json
@@ -2,15 +2,12 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "outDir": "ts",
-    "plugins": [
-      {
-        "name": "typescript-tslint-plugin",
-        "configFile": "./tslint.json",
-      }
-    ]
   },
   "include": [
     "src",
     "../../types"
+  ],
+  "exclude": [
+    "src/**/*-test.tsx"
   ]
 }
