+++ packages/styleguide/package.json
@@ -11,8 +11,8 @@
   "dependencies": {
     "@babel/polyfill": "7.2.5",
     "diffable-html": "4.0.0",
-    "pulsar-ui": "^0.1.0",
-    "pulsar-ui-theme-default": "^0.1.0",
+    "@pulsar-ui/core": "^0.1.0",
+    "@pulsar-ui/theme-default": "^0.1.0",
     "parse-prop-types": "0.3.0",
     "react-helmet": "5.2.0",
     "react-router-dom": "4.3.1",
