+++ packages/styleguide/src/Wrapper.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { ThemeProvider } from "pulsar-ui";
-import Theme from "pulsar-ui-theme-default";
+import { ThemeProvider } from "@pulsar-ui/core";
+import Theme from "@pulsar-ui/theme-default";
 
 const Wrapper = (props: any) => {
   return <ThemeProvider theme={Theme}>{props.children}</ThemeProvider>;
