+++ packages/theme-default/package.json
@@ -1,26 +1,27 @@
 {
-  "name": "@pulsar-ui/theme-blue",
+  "name": "@pulsar-ui/theme-default",
   "version": "0.1.0",
   "description": "",
-  "main": "build/index.js",
+  "main": "lib/index.js",
+  "module": "es/index.js",
+  "unpkg": "dist/pulsar-ui-theme-default.min.js",
+  "types": "ts",
   "scripts": {
-    "build": "ts-node index.ts",
+    "prebuild": "tsc --emitDeclarationOnly",
+    "start": "rollup -cw",
+    "build": "rollup -c",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/color": "^3.0.0",
-    "@types/mkdirp": "^0.5.2",
-    "@types/node": "^10.12.23",
-    "chroma-js": "^2.0.2",
-    "color": "^3.1.0",
-    "mkdirp": "^0.5.1",
-    "polished": "^2.3.3",
-    "ts-node": "^8.0.2",
+    "rollup": "1.2.2",
     "typescript": "3.2.4"
   },
+  "peerDependencies": {
+    "@pulsar-ui/core": "*"
+  },
   "dependencies": {
-    "pulsar-ui-theme-default": "^0.1.0"
+    "styled-tools": "1.7.1"
   }
 }
