--- packages/theme-default/scripts/compiler.ts
@@ -1,35 +1,0 @@
-import * as ts from "typescript";
-
-function compile(fileNames: string[], options: ts.CompilerOptions): void {
-  const program = ts.createProgram(fileNames, options);
-  const emitResult = program.emit();
-
-  const allDiagnostics = ts
-    .getPreEmitDiagnostics(program)
-    .concat(emitResult.diagnostics);
-
-  allDiagnostics.forEach(diagnostic => {
-    if (diagnostic.file) {
-      const { line, character } = diagnostic.file.getLineAndCharacterOfPosition(
-        diagnostic.start!
-      );
-      const message = ts.flattenDiagnosticMessageText(
-        diagnostic.messageText,
-        "\n"
-      );
-      process.stdout.write(
-        `${diagnostic.file.fileName} (${line + 1},${character + 1}): ${message}`
-      );
-    } else {
-      process.stdout.write(
-        `${ts.flattenDiagnosticMessageText(diagnostic.messageText, "\n")}`
-      );
-    }
-  });
-
-  const exitCode = emitResult.emitSkipped ? 1 : 0;
-  process.stdout.write(`Process exiting with code '${exitCode}'.`);
-  process.exit(exitCode);
-}
-
-export default compile;
