+++ packages/theme-default/src/elements/Avatar/Avatar.ts
@@ -1,0 +1,7 @@
+import { css } from "@pulsar-ui/core";
+
+const Avatar = css`
+  border-radius: 100%;
+`;
+
+export default Avatar;
