+++ packages/theme-default/src/elements/Avatar/index.ts
@@ -1,0 +1,2 @@
+import Avatar from "./Avatar";
+export default Avatar;
