+++ packages/theme-default/src/elements/Input/Input.ts
@@ -1,4 +1,4 @@
-import { css } from "pulsar-ui";
+import { css } from "@pulsar-ui/core";
 import InputRange from "./InputRange";
 import InputText from "./InputText";
