+++ packages/theme-default/src/elements/Input/InputText.ts
@@ -1,4 +1,4 @@
-import { css } from "pulsar-ui";
+import { css } from "@pulsar-ui/core";
 import { palette as p, theme } from "styled-tools";
 
 const InputText = css`
