+++ packages/theme-default/src/elements/index.ts
@@ -1,5 +1,7 @@
 import Input from "./Input";
+import Avatar from "./Avatar";
 
 export default {
-  Input
+  Input,
+  Avatar
 };
